Moved base.py into gaetest_common.py
