+++ tests/gaetest_common.py
@@ -11,6 +11,12 @@ BASE_DIR = os.path.abspath(os.path.join(my_dir, '..'))
 if BASE_DIR not in sys.path:
     sys.path.insert(0, BASE_DIR)
 
+from unittest import TestCase
+
+from google.appengine.ext import ndb, testbed
+from google.appengine.datastore import datastore_stub_util
+
+
 SAMPLE_AUTHORS = (
     ('Bob', 'Boston'),
     ('Harry', 'Houston'),
@@ -39,3 +45,34 @@ def fill_authors(Author):
         authors.append(author)
         AGE_BASE += 1
     return authors
+
+
+class NDBTestCase(TestCase):
+    def setUp(self):
+        self.testbed = testbed.Testbed()
+        self.testbed.activate()
+
+        policy = datastore_stub_util.PseudoRandomHRConsistencyPolicy(
+            probability=1)
+        self.testbed.init_datastore_v3_stub(consistency_policy=policy)
+
+        ctx = ndb.get_context()
+        ctx.set_cache_policy(False)
+        ctx.set_memcache_policy(False)
+
+    def tearDown(self):
+        self.testbed.deactivate()
+
+
+
+class DBTestCase(TestCase):
+    def setUp(self):
+        self.testbed = testbed.Testbed()
+        self.testbed.activate()
+
+        policy = datastore_stub_util.PseudoRandomHRConsistencyPolicy(
+            probability=1)
+        self.testbed.init_datastore_v3_stub(consistency_policy=policy)
+
+    def tearDown(self):
+        self.testbed.deactivate()
